Created a new core markdown extension folder. Moved preview links extension.
