+++ testproject/testproject/settings/__init__.py
@@ -103,7 +103,6 @@ INSTALLED_APPS = [
     'wiki.plugins.images',
     'wiki.plugins.attachments',
     'wiki.plugins.notifications',
-    'wiki.plugins.preview',
     'mptt',
 ]
 from django import VERSION
